# server/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "File Storage API",
    "version": "1.0.0",
    "description": "API for file upload, download, update and delete"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://app-e7khm2xacq-uc.a.run.app"
    }
  ],
  "tags": [
    {
      "name": "Files",
      "description": "File storage operations"
    }
  ],
  "paths": {
    "/api/files": {
      "get": {
        "tags": ["Files"],
        "summary": "List all files",
        "responses": {
          "200": {
            "description": "List of files",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/FileMeta" }
                }
              }
            }
          }
        }
      }
    },
    "/api/files/upload": {
      "post": {
        "tags": ["Files"],
        "summary": "Upload a file",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File uploaded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FileMeta" }
              }
            }
          }
        }
      }
    },
    "/api/files/download/{id}": {
      "get": {
        "tags": ["Files"],
        "summary": "Download file",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "File stream"
          }
        }
      }
    },
    "/api/files/content/{id}": {
      "get": {
        "tags": ["Files"],
        "summary": "Read file content",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Text content",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/{id}": {
      "delete": {
        "tags": ["Files"],
        "summary": "Delete file",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          }
        }
      }
    },
    "/api/firestore/{collectionName}": {
      "get": {
        "tags": ["Firestore"],
        "summary": "Get all user documents from collection",
        "parameters": [
          {
            "name": "collectionName",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "User filtered documents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        }
      },

      "post": {
        "tags": ["Firestore"],
        "summary": "Bulk insert documents",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bulk insert success"
          }
        }
      },

      "delete": {
        "tags": ["Firestore"],
        "summary": "Delete all user documents from collection",
        "parameters": [
          {
            "name": "collectionName",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "All docs deleted"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "FileMeta": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "mimeType": { "type": "string" },
          "size": { "type": "integer" },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}